Improved the module initialization design
