--- Nefarius.DSharpPlus.Extensions.Hosting/DiscordExtensionConfiguration.cs
@@ -1,8 +1,0 @@
-﻿namespace Nefarius.DSharpPlus.Extensions.Hosting;
-
-/// <summary>
-///     Implements configuration for optional extensions.
-/// </summary>
-public interface IDiscordExtensionConfiguration
-{
-}
