+++ Nefarius.DSharpPlus.SlashCommands.Extensions.Hosting/DiscordServiceCollectionExtensions.cs
@@ -1,9 +1,9 @@
-﻿using System;
+﻿#nullable enable
+
+using System;
 using System.Diagnostics.CodeAnalysis;
 
-using DSharpPlus;
 using DSharpPlus.SlashCommands;
-using DSharpPlus.SlashCommands.EventArgs;
 
 using Microsoft.Extensions.DependencyInjection;
 
@@ -11,7 +11,6 @@ using Nefarius.DSharpPlus.Extensions.Hosting;
 using Nefarius.DSharpPlus.Extensions.Hosting.Util;
 using Nefarius.DSharpPlus.SlashCommands.Extensions.Hosting.Attributes;
 using Nefarius.DSharpPlus.SlashCommands.Extensions.Hosting.Events;
-using Nefarius.DSharpPlus.SlashCommands.Extensions.Hosting.Util;
 
 namespace Nefarius.DSharpPlus.SlashCommands.Extensions.Hosting;
 
@@ -21,6 +20,7 @@ namespace Nefarius.DSharpPlus.SlashCommands.Extensions.Hosting;
 [SuppressMessage("ReSharper", "UnusedMember.Global")]
 [SuppressMessage("ReSharper", "ClassNeverInstantiated.Global")]
 [SuppressMessage("ReSharper", "MemberCanBePrivate.Global")]
+[SuppressMessage("ReSharper", "UnusedType.Global")]
 public static class DiscordServiceCollectionExtensions
 {
     /// <summary>
@@ -36,80 +36,13 @@ public static class DiscordServiceCollectionExtensions
     /// <returns>The <see cref="IServiceCollection" />.</returns>
     public static IServiceCollection AddDiscordSlashCommands(
         this IServiceCollection services,
-        Action<SlashCommandsConfiguration?> configuration = null,
-        Action<SlashCommandsExtension?> extension = null,
+        Action<SlashCommandsConfiguration>? configuration = null,
+        Action<SlashCommandsExtension>? extension = null,
         bool autoRegisterSubscribers = true
     )
     {
-        services.AddSingleton(typeof(IDiscordExtensionConfiguration), provider =>
-        {
-            SlashCommandsConfiguration options = new();
-
-            configuration?.Invoke(options);
-
-            //
-            // Make all services available to bot commands
-            // 
-            options.Services = provider;
-
-            DiscordClient discord = provider.GetRequiredService<IDiscordClientService>().Client;
-
-            SlashCommandsExtension ext = discord.UseSlashCommands(options);
-
-            extension?.Invoke(ext);
-
-            ext.ContextMenuErrored += async delegate(SlashCommandsExtension sender, ContextMenuErrorEventArgs args)
-            {
-                using IServiceScope scope = provider.CreateScope();
-
-                foreach (IDiscordSlashCommandsEventsSubscriber eventsSubscriber in scope
-                             .GetDiscordSlashCommandsEventsSubscriber())
-                {
-                    await eventsSubscriber.SlashCommandsOnContextMenuErrored(sender, args);
-                }
-            };
-
-            ext.ContextMenuExecuted +=
-                async delegate(SlashCommandsExtension sender, ContextMenuExecutedEventArgs args)
-                {
-                    using IServiceScope scope = provider.CreateScope();
-
-                    foreach (IDiscordSlashCommandsEventsSubscriber eventsSubscriber in scope
-                                 .GetDiscordSlashCommandsEventsSubscriber())
-                    {
-                        await eventsSubscriber.SlashCommandsOnContextMenuExecuted(sender, args);
-                    }
-                };
-
-            ext.SlashCommandErrored +=
-                async delegate(SlashCommandsExtension sender, SlashCommandErrorEventArgs args)
-                {
-                    using IServiceScope scope = provider.CreateScope();
-
-                    foreach (IDiscordSlashCommandsEventsSubscriber eventsSubscriber in scope
-                                 .GetDiscordSlashCommandsEventsSubscriber())
-                    {
-                        await eventsSubscriber.SlashCommandsOnSlashCommandErrored(sender, args);
-                    }
-                };
-
-            ext.SlashCommandExecuted +=
-                async delegate(SlashCommandsExtension sender, SlashCommandExecutedEventArgs args)
-                {
-                    using IServiceScope scope = provider.CreateScope();
-
-                    foreach (IDiscordSlashCommandsEventsSubscriber eventsSubscriber in scope
-                                 .GetDiscordSlashCommandsEventsSubscriber())
-                    {
-                        await eventsSubscriber.SlashCommandsOnSlashCommandExecuted(sender, args);
-                    }
-                };
-
-            //
-            // This is intentional; we don't need this "service", just the execution flow ;)
-            // 
-            return null;
-        });
+        services.AddSingleton<IServiceActivator, SlashCommandsActivator>(_ =>
+            new SlashCommandsActivator(configuration, extension));
 
         if (!autoRegisterSubscribers)
         {
